Calculate carbon statistics

These functions allow to calculated statistics based on the harmonized carbon layers for 2010 and 2018 by Noon et al. (2022).

Usage

calc_irr_carbon(
  type = c("total", "soil", "biomass", "all"),
  engine = "extract",
  stats = "mean"
)

calc_man_carbon(
  type = c("total", "soil", "biomass", "all"),
  engine = "extract",
  stats = "mean"
)

calc_vul_carbon(
  type = c("total", "soil", "biomass", "all"),
  engine = "extract",
  stats = "mean"
)

Arguments

type

One of "total", "soil", "biomass", "all". Determines for which data layer the statistics are calculated.

engine

The preferred processing functions from either one of "zonal", "extract" or "exactextract" as character.

stats

Function to be applied to compute statistics for polygons either one or multiple inputs as character. Supported statistics are: "mean", "median", "sd", "min", "max", "sum", and "var".

Value

A function that returns a tibble with a column for each statistic and rows for every year and type of carbon.

Details

The required resources for these indicators are:

carbon_resources

Irrecoverable carbon is the amount of carbon that, if lost today, could not be recovered until 2050. It can be calculated for above- and below-ground carbon, the total amount of carbon, or for all layers.

Manageable carbon is the amount of carbon that, in principle, is manageable by human activities, e.g. its release to the atmosphere can be prevented. It can be calculated for above- and below-ground carbon, the total amount of carbon, or for all layers.

Vulnerable carbon is the amount of carbon that would be released in a typical land conversion activity. It can be calculated for above- and below-ground carbon, the total amount of carbon, or for all layers.